// src/client/models/UserCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserCreate = {
    name: string;
    /**
     * Custom validated email address
     */
    email: string;
    role: string;
    is_active: boolean;
    password: string;
};