// src/client/models/UserUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserUpdate = {
    id: number;
    name: (string | null);
    email: (string | null);
    role: (string | null);
    password?: (string | null);
    is_active: (boolean | null);
};