// src/client/models/UserOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserOut = {
    name: string;
    /**
     * Custom validated email address
     */
    email: string;
    role: string;
    is_active: boolean;
    id: number;
    created_at: string;
    updated_at: string;
};